% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCluster.R
\name{createCluster}
\alias{createCluster}
\title{Create clusters}
\usage{
createCluster(inputDir, outputDir, finaloutput = FALSE)
}
\arguments{
\item{inputDir}{String}

\item{outputDir}{String}

\item{finaloutput}{Boolean}
}
\value{
writes a text file named as OneMimatchCluster.txt
}
\description{
Creates clusters of parents and one-mismatch children.
}
