% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanAndSD.R
\name{getMeanAndSD}
\alias{getMeanAndSD}
\title{Properties of clusters}
\usage{
getMeanAndSD(clusters)
}
\arguments{
\item{clusters}{A table of clusters with parents and children}
}
\description{
Mean and standard deviation for each cluster are calculated and then means and
standard deviations are fitted to a LOESS regression as a function of parent abundance.
}
