% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer}
\alias{infer}
\title{Inference test}
\usage{
infer(clusterProperties, cluster, abundanceThreshold, outputDir)
}
\arguments{
\item{clusterProperties}{A table that includes the properties of the clusters}

\item{cluster}{a cluster table}

\item{abundanceThreshold}{threshold for removing low abundant sequences}

\item{outputDir}{output directory}
}
\value{
a table with significant children
}
\description{
Inference test
}
