% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCluster.R
\name{loadCluster}
\alias{loadCluster}
\title{Loading the cluster table "OneMismatchCluster.txt"}
\usage{
loadCluster(outputDir)
}
\arguments{
\item{outputDir}{Output Directory where OneMismatchCluster.txt is placed}
}
\value{

}
\description{
Loading the cluster table "OneMismatchCluster.txt"
}
