% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processClusters.R
\name{processCluster}
\alias{processCluster}
\title{Create clusters}
\usage{
processCluster(inputDir, outputDir, countTable = FALSE)
}
\arguments{
\item{inputDir}{String}

\item{outputDir}{String}

\item{countTable}{Boolean}
}
\value{
writes a text file named as OneMimatchCluster.txt
}
\description{
Creates clusters of parents and one-mismatch children.
}
