% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFastaSeq.R
\name{writeFastaSeq}
\alias{writeFastaSeq}
\title{Fasta file including inferred sequences and parents}
\usage{
writeFastaSeq(inferredseq, parents, outputDir)
}
\arguments{
\item{outputDir}{}
}
\value{

}
\description{
Fasta file including inferred sequences and parents
}
