% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFastaSequences.R
\name{writeInferedSequences}
\alias{writeInferedSequences}
\title{Fasta file including inferred sequences and parents}
\usage{
writeInferedSequences(inferredseq, parents, outputDir)
}
\arguments{
\item{outputDir}{}
}
\value{

}
\description{
Fasta file including inferred sequences and parents
}
