% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAbundantParents.R
\name{getAbundantParents}
\alias{getAbundantParents}
\title{Parents with abundance more than a threshold}
\usage{
getAbundantParents(cluster, abundanceThreshold = 1000)
}
\arguments{
\item{cluster}{the table loaded by loadCluster method}

\item{abundanceThreshold}{threshold for filtering sequences (defualt=1000)}
}
\value{
a table with parents having abundance above threshold
}
\description{
Selects only parents that have an abundance above a threshold
}
