% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTrueSequences.R
\name{inferTrueSequences}
\alias{inferTrueSequences}
\title{Infer true amplicon sequence variants}
\usage{
inferTrueSequences(
  inputDir,
  outputDir,
  abundanceThreshold = 1000,
  countTable = FALSE
)
}
\arguments{
\item{inputDir}{String}

\item{outputDir}{String}

\item{countTable}{Boolean}
}
\value{
TODO:  determine what result to return
}
\description{
Invokes algorithm to infer true amplicon sequence variants.
}
