% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLoessPlot.R
\name{makeLoessPlot}
\alias{makeLoessPlot}
\title{Relationship between parent abundance and mean and standard deviation}
\usage{
makeLoessPlot(clusterProperties, param)
}
\arguments{
\item{param}{"mean" or "sd"}
}
\value{

}
\description{
Relationship between parent abundance and mean and standard deviation
}
