% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLoessPlot.R
\name{makeLoessPlot}
\alias{makeLoessPlot}
\title{Plot showing the relationship between parent abundance and mean and standard deviation of one-mismatch children.}
\usage{
makeLoessPlot(outputDir, fileName = "childrenProperties.txt")
}
\arguments{
\item{outputDir}{The directory where childrenproperties.txt has been saved}

\item{fileName}{Default=childrenproperties.txt}
}
\value{
Figures
}
\description{
This plot helps to choose an abundance threshold for the "inferTrueSequence.R" function.
Based on this plot, a threshold at which the mean or standard deviation of one-mismatch children are
not a smooth function of the parent abundance can be chosen for filtering the low abundant sequence variants.
}
