% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLoessPlot.R
\name{makeLoessPlot}
\alias{makeLoessPlot}
\title{Plot showing the relationship between parent abundance and mean and standard deviation}
\usage{
makeLoessPlot(outputDir, fileName = "childrenproperties.txt")
}
\arguments{
\item{outputDir}{The directory where childrenproperties.txt has been saved}

\item{fileName}{Default=childrenproperties.txt}
}
\value{
Figures
}
\description{
Plot showing the relationship between parent abundance and mean and standard deviation
}
