% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLoessPlot.R
\name{makeLoessPlot}
\alias{makeLoessPlot}
\title{Plot showing the relationship between parent abundance and mean and standard deviation}
\usage{
makeLoessPlot(clusterProperties)
}
\arguments{
\item{clusterProperties}{a table of clusters}
}
\value{
Figures
}
\description{
Plot showing the relationship between parent abundance and mean and standard deviation
}
