% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFastaSeq.R
\name{writeFastaSeq}
\alias{writeFastaSeq}
\title{Fasta file including inferred sequences and parents}
\usage{
writeFastaSeq(inferredseq, parents, outputDir)
}
\arguments{
\item{inferredseq}{Significant children}

\item{parents}{parents with abundance above the threshold}

\item{outputDir}{output directory}
}
\value{
writes a fasta file
}
\description{
Fasta file including inferred sequences and parents
}
